+++ lib/screens/home.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:gamelib_mob/list/list_class.dart';
+import 'package:gamelib_mob/list/list_item.dart';
 import 'package:gamelib_mob/list/main_list.dart';
 import 'package:gamelib_mob/screens/search_game.dart';
 import 'package:gamelib_mob/screens/game_detail.dart';
